+++ src/engine/i_system.c
@@ -74,7 +74,7 @@ static int64    I_GetTime_Scale = 1 << 24;
 // I_uSleep
 //
 
-void I_Sleep(uint32_t usecs) {
+void I_Sleep(long usecs) {
 	SDL_Delay(usecs);
 }
 
@@ -171,7 +171,7 @@ void I_EndDisplay(void) {
 //
 
 fixed_t I_GetTimeFrac(void) {
-	uint32_t now;
+	long now;
 	fixed_t frac;
 
 	now = SDL_GetTicks();
@@ -369,7 +369,7 @@ int I_GetTimeMS(void) {
 // I_GetRandomTimeSeed
 //
 
-uint32_t I_GetRandomTimeSeed(void) {
+unsigned long I_GetRandomTimeSeed(void) {
 	// not exactly random....
 	return SDL_GetTicks();
 }
